@@ -4,6 +4,8 @@ import io.github.compactrisc16.assembler.isa.instruction.AbstractInstruction;
 import io.github.compactrisc16.assembler.isa.instruction.exception.InstructionParseException;
 import io.github.compactrisc16.assembler.isa.register.Register;
 
+import java.util.List;
+
 /**
  * {@link RdestImmInstruction} is an {@link AbstractInstruction} with <code>Rdest</code> and <code>Imm</code> as
  * arguments.
@@ -27,21 +29,29 @@ public class RdestImmInstruction extends AbstractInstruction {
     }
 
     @Override
-    public void parse(String[] assemblyInstruction) throws InstructionParseException {
-        if (assemblyInstruction.length != 3) {
+    public void parse(List<String> lineWords) throws InstructionParseException {
+        if (lineWords.size() != 3) {
             throw new InstructionParseException(
                     String.format("Invalid arguments. Expected: %s <Rdest>, <Imm>", mnemonic));
         }
 
-        rdest = parseRegister(assemblyInstruction[INSTRUCTION_INDEX_RDEST]);
+        rdest = parseRegister(lineWords.get(INSTRUCTION_INDEX_RDEST));
         // Note that the maximum allowable value is '0xFF' because the assembly can contain unsigned integers, and it's
         // up to the programmer to determine if the number should be interpreted as unsigned or signed numbers at
         // CR16 runtime.
-        imm = parseImmediate(assemblyInstruction[INSTRUCTION_INDEX_IMM], Byte.MIN_VALUE, 0xFF);
+        imm = parseImmediate(lineWords.get(INSTRUCTION_INDEX_IMM), Byte.MIN_VALUE, 0xFF);
     }
 
     @Override
     public int assemble() {
         return super.assemble() | rdest.getIndex() << 8 | (imm & 0xFF);
+    }
+
+    public Register getRdest() {
+        return rdest;
+    }
+
+    public int getImm() {
+        return imm;
     }
 }
