@@ -4,6 +4,8 @@ import io.github.compactrisc16.assembler.isa.instruction.exception.InstructionPa
 import io.github.compactrisc16.assembler.isa.instruction.type.OpcodeExtInstruction;
 import io.github.compactrisc16.assembler.isa.register.Register;
 
+import java.util.List;
+
 /**
  * {@link CallInstruction} is an {@link OpcodeExtInstruction} representing a <code>CALL</code> instruction with
  * <code>Rtarget</code> as an argument.
@@ -22,16 +24,20 @@ public class CallInstruction extends OpcodeExtInstruction {
     }
 
     @Override
-    public void parse(String[] assemblyInstruction) throws InstructionParseException {
-        if (assemblyInstruction.length != 2) {
+    public void parse(List<String> lineWords) throws InstructionParseException {
+        if (lineWords.size() != 2) {
             throw new InstructionParseException(String.format("Invalid arguments. Expected: %s <Rtarget>", mnemonic));
         }
 
-        rtarget = parseRegister(assemblyInstruction[INSTRUCTION_INDEX_RTARGET]);
+        rtarget = parseRegister(lineWords.get(INSTRUCTION_INDEX_RTARGET));
     }
 
     @Override
     public int assemble() {
         return super.assemble() | rtarget.getIndex();
+    }
+
+    public Register getRtarget() {
+        return rtarget;
     }
 }
