@@ -4,6 +4,8 @@ import io.github.compactrisc16.assembler.isa.instruction.exception.InstructionPa
 import io.github.compactrisc16.assembler.isa.instruction.type.OpcodeExtInstruction;
 import io.github.compactrisc16.assembler.isa.register.Register;
 
+import java.util.List;
+
 /**
  * {@link RdestInstruction} is an {@link OpcodeExtInstruction} with <code>Rdest</code> as an argument.
  */
@@ -25,16 +27,20 @@ public class RdestInstruction extends OpcodeExtInstruction {
     }
 
     @Override
-    public void parse(String[] assemblyInstruction) throws InstructionParseException {
-        if (assemblyInstruction.length != 2) {
+    public void parse(List<String> lineWords) throws InstructionParseException {
+        if (lineWords.size() != 2) {
             throw new InstructionParseException(String.format("Invalid arguments. Expected: %s <Rdest>", mnemonic));
         }
 
-        rdest = parseRegister(assemblyInstruction[INSTRUCTION_INDEX_RDEST]);
+        rdest = parseRegister(lineWords.get(INSTRUCTION_INDEX_RDEST));
     }
 
     @Override
     public int assemble() {
         return super.assemble() | rdest.getIndex() << 8;
+    }
+
+    public Register getRdest() {
+        return rdest;
     }
 }
